requirements.txt
accelerate
controlnet_aux
diffusers==0.14.0
gdown
huggingface
hydra-core
jupyter
mediapipe==0.10.0
matplotlib
numpy
opencv-contrib-python
pandas
pillow
pyiqa
pycocotools
tqdm
transformers
ultralytics
wandb
wget
xformers
